# docs/swagger.yaml
info:
  contact: {}
  license: {}
paths:
  /api/v1/articles:
    get:
      parameters:
      - description: 状态
        in: query
        name: state
        type: integer
      - description: 标签id
        in: query
        name: tag_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取文章
      tags:
      - article文章管理
    post:
      parameters:
      - description: 标题
        in: query
        name: title
        required: true
        type: string
      - description: 描述
        in: query
        name: desc
        required: true
        type: string
      - description: 内容
        in: query
        name: content
        required: true
        type: string
      - description: 创建者
        in: query
        name: created_by
        required: true
        type: string
      - description: 标签id
        in: query
        name: tag_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增文章
      tags:
      - article文章管理
  /api/v1/articles/{id}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除文章
      tags:
      - article文章管理
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 根据id获取文章
      tags:
      - article文章管理
    put:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: 标题
        in: query
        name: title
        required: true
        type: string
      - description: 描述
        in: query
        name: desc
        required: true
        type: string
      - description: 内容
        in: query
        name: content
        required: true
        type: string
      - description: 修改者
        in: query
        name: modified_by
        required: true
        type: string
      - description: 标签id
        in: query
        name: tag_id
        required: true
        type: integer
      - description: 标签状态
        in: query
        name: state
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 修改文章
      tags:
      - article文章管理
  /api/v1/tags:
    get:
      parameters:
      - description: 名称
        in: query
        name: name
        type: string
      - description: 状态
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取文章标签
      tags:
      - tag标签管理
    post:
      parameters:
      - description: Name
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: CreatedBy
        in: query
        name: created_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增文章标签
      tags:
      - tag标签管理
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增文章标签
      tags:
      - tag标签管理
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Name
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: ModifiedBy
        in: query
        name: modified_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 修改文章标签
      tags:
      - tag标签管理
  /auth:
    get:
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取token
      tags:
      - auth管理
swagger: "2.0"